backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
